+++ .appGenInfo.json
@@ -1,22 +1,22 @@
 {
   "generationParameters": {
-    "generationDate": "Mon Aug 18 2025 06:46:17 GMT+0000 (Coordinated Universal Time)",
+    "generationDate": "Wed Aug 20 2025 00:50:02 GMT+0000 (Coordinated Universal Time)",
     "generatorPlatform": "SAP Business Application Studio",
-    "serviceType": "None",
-    "metadataFilename": "",
-    "serviceUrl": "N/A",
-    "appName": "casestudy",
-    "appTitle": "Case Study",
-    "appDescription": "An SAP Fiori application.",
-    "appNamespace": "",
+    "serviceType": "File",
+    "metadataFilename": "metadata.xml",
+    "serviceUrl": "/here/goes/your/serviceurl/",
+    "appName": "order_management",
+    "appTitle": "Order Management",
+    "appDescription": "An SAP Fiori application for case study.",
+    "appNamespace": "com.order_management",
     "ui5Theme": "sap_horizon",
     "ui5Version": "1.139.0",
     "enableCodeAssist": false,
     "enableEslint": false,
     "enableTypeScript": false,
-    "showMockDataInfo": false,
+    "showMockDataInfo": true,
     "generatorVersion": "1.18.5",
-    "template": "Basic",
+    "template": "Basic V2",
     "generatorName": "SAP Fiori Application Generator",
     "entityRelatedConfig": [],
     "launchText": "To launch the generated application, run the following from the generated application root folder:\n\n```\n    npm start\n```"
