--- webapp/fragment/ValueHelpDialog.fragment.xml
@@ -1,9 +1,0 @@
-<core:FragmentDefinition xmlns="sap.m"
-    xmlns:core="sap.ui.core">
-    <Dialog id="idProductDialog" title="Test">
-        <beginButton>
-            <Button id="idButtonCloseDialog" text="Close" press="onCloseDialog"/>
-        </beginButton>
-    </Dialog>
-
-</core:FragmentDefinition>
