+++ webapp/i18n/i18n.properties
@@ -1,18 +1,11 @@
-# This is the resource bundle for casestudy
+# This is the resource bundle for com.ordermanagement.ordermanagement
 
 #Texts for manifest.json
 
 #XTIT: Application name
-appTitle=Case Study
+appTitle=Order Management
 
 #YDES: Application description
-appDescription=An SAP Fiori application.
+appDescription=An SAP Fiori application for case study.
 #XTIT: Main view title
-title=Case Study
-
-#XTIT: Create Order
-pageCreateOrder=Create Order
-
-#XPAN: Order Details
-lblReceivingPlant=Receiving Plant
-lbldeliveringPlant=Delivering Plant
+title=Order Management
