+++ webapp/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Case Study</title>
+    <title>Order Management</title>
     <style>
         html, body, body > div, #container, #container-uiarea {
             height: 100%;
@@ -15,7 +15,7 @@
         src="resources/sap-ui-core.js"
         data-sap-ui-theme="sap_horizon"
         data-sap-ui-resource-roots='{
-            "casestudy": "./"
+            "com.ordermanagement.ordermanagement": "./"
         }'
         data-sap-ui-on-init="module:sap/ui/core/ComponentSupport"
         data-sap-ui-compat-version="edge"
@@ -26,9 +26,9 @@
 <body class="sapUiBody sapUiSizeCompact" id="content">
     <div
         data-sap-ui-component
-        data-name="casestudy"
+        data-name="com.ordermanagement.ordermanagement"
         data-id="container"
-        data-settings='{"id" : "casestudy"}'
+        data-settings='{"id" : "com.ordermanagement.ordermanagement"}'
         data-handle-validation="true"
     ></div>
 </body>
