+++ webapp/localService/mainService/data/Orders.json
@@ -1,0 +1,67 @@
+[
+  {
+    "OrderNumber": 1001,
+    "CreationDate": "/Date(1609459200000)/",
+    "ReceivingPlantCode": 101,
+    "ReceivingPlantDescription": "Plant A",
+    "DeliveringPlantCode": 201,
+    "DeliveringPlantDescription": "Depot A",
+    "Status": "Created",
+    "ProductID": 301,
+    "ProductName": "Laptop",
+    "Quantity": 3,
+    "PricePerQuantity": 500
+  },
+  {
+    "OrderNumber": 1002,
+    "CreationDate": "/Date(1612137600000)/",
+    "ReceivingPlantCode": 102,
+    "ReceivingPlantDescription": "Plant B",
+    "DeliveringPlantCode": 202,
+    "DeliveringPlantDescription": "Depot B",
+    "Status": "Released",
+    "ProductID": 302,
+    "ProductName": "Samsung TV",
+    "Quantity": 5,
+    "PricePerQuantity": 400
+  },
+  {
+    "OrderNumber": 1003,
+    "CreationDate": "/Date(1614556800000)/",
+    "ReceivingPlantCode": 103,
+    "ReceivingPlantDescription": "Plant C",
+    "DeliveringPlantCode": 203,
+    "DeliveringPlantDescription": "Depot C",
+    "Status": "Partially Completed",
+    "ProductID": 303,
+    "ProductName": "Bluetooth Speaker",
+    "Quantity": 2,
+    "PricePerQuantity": 150
+  },
+  {
+    "OrderNumber": 1004,
+    "CreationDate": "/Date(1617235200000)/",
+    "ReceivingPlantCode": 104,
+    "ReceivingPlantDescription": "Plant D",
+    "DeliveringPlantCode": 204,
+    "DeliveringPlantDescription": "Depot D",
+    "Status": "Delivered",
+    "ProductID": 304,
+    "ProductName": "Smartphone",
+    "Quantity": 4,
+    "PricePerQuantity": 300
+  },
+  {
+    "OrderNumber": 1005,
+    "CreationDate": "/Date(1619827200000)/",
+    "ReceivingPlantCode": 105,
+    "ReceivingPlantDescription": "Plant E",
+    "DeliveringPlantCode": 205,
+    "DeliveringPlantDescription": "Depot E",
+    "Status": "Created",
+    "ProductID": 305,
+    "ProductName": "Tablet",
+    "Quantity": 6,
+    "PricePerQuantity": 250
+  }
+]
