+++ webapp/localService/mainService/metadata.xml
@@ -1,0 +1,33 @@
+<edmx:Edmx Version="1.0" xmlns:edmx="http://schemas.microsoft.com/ado/2007/06/edmx" xmlns:sap="http://www.sap.com/Protocols/SAPData">
+    <edmx:DataServices xmlns:m="http://schemas.microsoft.com/ado/2007/08/dataservices/metadata" m:DataServiceVersion="2.0">
+        <Schema xmlns="http://schemas.microsoft.com/ado/2008/09/edm" Namespace="OrderModel">
+            <!-- Entity Definition -->
+            <EntityType Name="Order">
+
+                <!-- Orders -->
+                <Property Name="OrderNumber" Type="Edm.Int32" Nullable="false" sap:label="Order Number" />
+                <Property Name="CreationDate" Type="Edm.DateTime" Nullable="false" sap:label="Creation Date" />
+                <Property Name="ReceivingPlantCode" Type="Edm.Int32" Nullable="false" sap:label="Receiving Plant Code" />
+                <Property Name="ReceivingPlantDescription" Type="Edm.String" Nullable="false" sap:label="Receiving Plant Description" />
+                <Property Name="DeliveringPlantCode" Type="Edm.Int32" Nullable="false" sap:label="Delivering Plant Code" />
+                <Property Name="DeliveringPlantDescription" Type="Edm.String" Nullable="false" sap:label="Delivering Plant Description" />
+                <Property Name="Status" Type="Edm.String" Nullable="false" sap:label="Order Status" />
+
+                <!-- Product -->
+                 <Key>
+                    <PropertyRef Name="ProductID" />
+                </Key>
+                <Property Name="ProductID" Type="Edm.Int32" Nullable="false" sap:label="Product ID" />
+                <Property Name="ProductName" Type="Edm.String" Nullable="false" sap:label="Product Name" />
+                <Property Name="Quantity" Type="Edm.Int32" Nullable="false" sap:label="Quantity" />
+                <Property Name="PricePerQuantity" Type="Edm.Int32" Nullable="false" sap:label="Price Per Quantity" />
+                </EntityType>
+ 
+            <!-- Entity Container -->
+            <EntityContainer Name="OrderEntities" m:IsDefaultEntityContainer="true">
+                <EntitySet Name="Orders" EntityType="OrderModel.Order" />
+            </EntityContainer>
+        </Schema>
+    </edmx:DataServices>
+</edmx:Edmx>
+
