+++ webapp/manifest.json
@@ -1,7 +1,7 @@
 {
   "_version": "1.76.0",
   "sap.app": {
-    "id": "casestudy",
+    "id": "com.ordermanagement.ordermanagement",
     "type": "application",
     "i18n": "i18n/i18n.properties",
     "applicationVersion": {
@@ -13,7 +13,18 @@
     "sourceTemplate": {
       "id": "@sap/generator-fiori:basic",
       "version": "1.18.5",
-      "toolsId": "54157a07-fd31-4f35-814b-ff8afa3c3c94"
+      "toolsId": "e010e601-6ef6-44d0-b1d3-0955dccecc28"
+    },
+    "dataSources": {
+      "mainService": {
+        "uri": "/here/goes/your/serviceurl/",
+        "type": "OData",
+        "settings": {
+          "annotations": [],
+          "localUri": "localService/mainService/metadata.xml",
+          "odataVersion": "2.0"
+        }
+      }
     }
   },
   "sap.ui": {
@@ -49,8 +60,13 @@
       "i18n": {
         "type": "sap.ui.model.resource.ResourceModel",
         "settings": {
-          "bundleName": "casestudy.i18n.i18n"
+          "bundleName": "com.ordermanagement.ordermanagement.i18n.i18n"
         }
+      },
+      "": {
+        "dataSource": "mainService",
+        "preload": true,
+        "settings": {}
       }
     },
     "resources": {
@@ -68,28 +84,28 @@
         "transition": "slide",
         "type": "View",
         "viewType": "XML",
-        "path": "casestudy.view",
+        "path": "com.ordermanagement.ordermanagement.view",
         "async": true,
-        "viewPath": "casestudy.view"
+        "viewPath": "com.ordermanagement.ordermanagement.view"
       },
       "routes": [
         {
-          "name": "Routeg5_CaseStudy",
+          "name": "RouteOrderManagement",
           "pattern": ":?query:",
           "target": [
-            "Targetg5_CaseStudy"
+            "TargetOrderManagement"
           ]
         }
       ],
       "targets": {
-        "Targetg5_CaseStudy": {
-          "id": "g5_CaseStudy",
-          "name": "g5_CaseStudy"
+        "TargetOrderManagement": {
+          "id": "OrderManagement",
+          "name": "OrderManagement"
         }
       }
     },
     "rootView": {
-      "viewName": "casestudy.view.App",
+      "viewName": "com.ordermanagement.ordermanagement.view.App",
       "type": "XML",
       "id": "App",
       "async": true
