+++ webapp/test/integration/NavigationJourney.js
@@ -3,7 +3,7 @@
 sap.ui.define([
 	"sap/ui/test/opaQunit",
 	"./pages/App",
-	"./pages/g5_CaseStudy"
+	"./pages/OrderManagement"
 ], function (opaTest) {
 	"use strict";
