+++ webapp/test/integration/arrangements/Startup.js
@@ -14,7 +14,7 @@ sap.ui.define([
 			// start the app UI component
 			this.iStartMyUIComponent({
 				componentConfig: {
-					name: "casestudy",
+					name: "com.ordermanagement.ordermanagement",
 					async: true
 				},
 				hash: oOptions.hash,
