+++ webapp/test/integration/opaTests.qunit.html
@@ -9,7 +9,7 @@
 		src="../../resources/sap-ui-core.js"
 		data-sap-ui-theme="sap_horizon"
 		data-sap-ui-resource-roots='{
-			"casestudy": "../../",
+			"com.ordermanagement.ordermanagement": "../../",
 			"integration": "./"
 		}'
 		data-sap-ui-animation-mode="none"
