+++ webapp/test/integration/opaTests.qunit.js
@@ -1,7 +1,7 @@
 /* global QUnit */
 QUnit.config.autostart = false;
 
-sap.ui.require(["casestudy/test/integration/AllJourneys"
+sap.ui.require(["com/ordermanagement/ordermanagement/test/integration/AllJourneys"
 ], function () {
 	QUnit.start();
 });
