+++ webapp/test/integration/pages/OrderManagement.js
@@ -2,7 +2,7 @@ sap.ui.define([
 	"sap/ui/test/Opa5"
 ], function (Opa5) {
 	"use strict";
-	var sViewName = "g5_CaseStudy";
+	var sViewName = "OrderManagement";
 	
 	Opa5.createPageObjects({
 		onTheViewPage: {
