+++ webapp/test/unit/unitTests.qunit.html
@@ -2,12 +2,12 @@
 <html>
 <head>
 	<meta charset="utf-8">
-	<title>Unit tests for casestudy</title>
+	<title>Unit tests for com.ordermanagement.ordermanagement</title>
 	<script
 		id="sap-ui-bootstrap"
 		src="../../resources/sap-ui-core.js"
 		data-sap-ui-resource-roots='{
-			"casestudy": "../../",
+			"com/ordermanagement/ordermanagement": "../../",
 			"unit": "."
 		}'
 		data-sap-ui-async="true"
