+++ webapp/view/App.view.xml
@@ -1,4 +1,4 @@
-<mvc:View controllerName="casestudy.controller.App"
+<mvc:View controllerName="com.ordermanagement.ordermanagement.controller.App"
     displayBlock="true"
     xmlns:mvc="sap.ui.core.mvc"
     xmlns="sap.m">
