--- webapp/view/CreateOrder.view.xml
@@ -1,76 +1,0 @@
-<mvc:View xmlns:mvc="sap.ui.core.mvc" xmlns="sap.m" controllerName="casestudy.controller.CreateOrder">
-    <Page id="pageCreateOrder" title="{i18n>pageCreateOrder}" floatingFooter="true">
-    <!-- Panel: Order Details -->
-        <Panel xmlns="sap.m" width="auto" expandable="false" expanded="false" id="panOrderDetails" class="sapUiResponsiveMargin" headerText="OrderDetail">
-                <content>
-                    <HBox id="fbCreateOrder" justifyContent="Center" >
-                        <sap.ui.layout.form:SimpleForm xmlns:sap.ui.layout.form="sap.ui.layout.form" xmlns:sap.ui.core="sap.ui.core" editable="true" layout="ColumnLayout" id="formOrderDetails" width="100%" >
-                            <sap.ui.layout.form:content>
-                                <Label text="{i18n>lblReceivingPlant}" id="lblReceivingPlant" required="true"/>
-                                <Input id="inpReceivingPlant" width="70%" showValueHelp="true" valueHelpRequest=".onValueHelpRequest"/>                    
-                                <Label text="{i18n>lbldeliveringPlant}" id="lblDeliveringPlant" required="true"/>
-                                <Input id="inpDeliveringPlant" width="70%" showValueHelp="true" valueHelpRequest=".onValueHelpRequest"/>
-                            </sap.ui.layout.form:content>
-                        </sap.ui.layout.form:SimpleForm>
-                    </HBox>
-                </content>
-            </Panel>
-        <!-- Panel: Products -->
-        <Panel xmlns="sap.m" width="auto" expandable="false" expanded="false" id="panProduct" class="sapUiResponsiveMargin" headerText="Products">
-            <headerToolbar>
-                <Toolbar id="toolbar1">
-                    <Title id="ttlProd" text="Products(0)"/> <!--No Data Fed-->
-                    <ToolbarSpacer id="tsProduct"/>
-                    <Button id="btnAdd" icon="sap-icon://add"/>
-                    <Button id="btnDelete" icon="sap-icon://delete"/>
-                    </Toolbar>
-                
-            </headerToolbar>
-            <content>
-
-            <Table noDataText="No Data" id="tbProd" items="{/Products}"> <!--No Data fed-->
-                <items>            
-                    <ColumnListItem type="Active" id="colListProd">            
-                        <cells>
-                            <!-- <Text text="Row 1 Cell 1" id="text0"/>
-                            <Text text="Row 1 Cell 2" id="text1"/>
-                            <Text text="Row 1 Cell 3" id="text2"/> -->
-                        </cells>
-                    </ColumnListItem>
-                </items>
-                <columns>
-                    <Column id="colProdName">
-                        <header>
-                            <CheckBox id="cbProdName" text="Product Name"/>   
-                        </header>
-                    </Column>
-                    <Column id="colQnty">
-                        <header>
-                            <Label text="Quantity" id="lblQnty"/>
-                        </header>
-                    </Column>
-                    <Column id="colPriceQnty">
-                        <header>
-                            <Label text="Price for Quantity" id="lblPriceQnty"/>
-                        </header>
-                    </Column>
-                    <Column id="colTotalPrice">
-                        <header>
-                        <Label text="Total Price" id="lblTotalPrice"/>
-                        </header>
-                    </Column>
-                </columns>
-                </Table>
-            </content>
-            </Panel>
-    <content/>
-    <footer>
-        <OverflowToolbar id="idOFTFooter">
-        <ToolbarSpacer id="idTSpacerFooter"/>
-            <Button id="btnSave" text="Save" type="Emphasized" tooltip="Save"/>
-            <Button id="btnCancel" text="Cancel" type="Reject" tooltip="Cancel"/>
-        <content/>
-        </OverflowToolbar>
-    </footer>
-    </Page>
-</mvc:View>
