--- webapp/view/g5_CaseStudy.view.xml
@@ -1,6 +1,0 @@
-<mvc:View controllerName="casestudy.controller.g5_CaseStudy"
-    xmlns:mvc="sap.ui.core.mvc"
-    xmlns="sap.m">
-    <Page id="page" title="{i18n>title}">
-    </Page>
-</mvc:View>
